+++ ethereum/statetransition/src/main/java/tech/pegasys/teku/statetransition/OperationAddedSubscriber.java
@@ -1,5 +1,5 @@
 /*
- * Copyright ConsenSys Software Inc., 2022
+ * Copyright ConsenSys Software Inc., 2023
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -11,14 +11,11 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.beacon.sync;
+package tech.pegasys.teku.statetransition;
 
-import tech.pegasys.teku.infrastructure.events.EventChannels;
+import tech.pegasys.teku.statetransition.validation.InternalValidationResult;
 
-/**
- * CAUTION: this API is unstable and primarily intended for debugging and testing purposes this API
- * might be changed in any version in backward incompatible way
- */
-public interface SyncServiceFactory {
-  SyncService create(EventChannels eventChannels);
+public interface OperationAddedSubscriber<T> {
+  void onOperationAdded(
+      T operation, InternalValidationResult validationStatus, boolean fromNetwork);
 }