+++ networking/eth2/src/main/java/tech/pegasys/teku/networking/eth2/rpc/beaconchain/BeaconChainMethods.java
@@ -21,7 +21,6 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
-import org.apache.tuweni.bytes.Bytes;
 import org.hyperledger.besu.plugin.services.MetricsSystem;
 import tech.pegasys.teku.infrastructure.async.AsyncRunner;
 import tech.pegasys.teku.infrastructure.bytes.Bytes4;
@@ -48,7 +47,6 @@ import tech.pegasys.teku.networking.p2p.rpc.RpcMethod;
 import tech.pegasys.teku.spec.Spec;
 import tech.pegasys.teku.spec.SpecMilestone;
 import tech.pegasys.teku.spec.datastructures.blocks.SignedBeaconBlock;
-import tech.pegasys.teku.spec.datastructures.blocks.SignedBeaconBlockSchema;
 import tech.pegasys.teku.spec.datastructures.blocks.blockbody.versions.deneb.SignedBeaconBlockAndBlobsSidecar;
 import tech.pegasys.teku.spec.datastructures.execution.versions.deneb.BlobsSidecar;
 import tech.pegasys.teku.spec.datastructures.networking.libp2p.rpc.BeaconBlockAndBlobsSidecarByRootRequestMessage;
@@ -207,48 +205,25 @@ public class BeaconChainMethods {
         BeaconBlocksByRootRequestMessage.SSZ_SCHEMA;
     final boolean expectResponseToRequest = true;
 
-    // V1 request only deal with Phase0 blocks
-    final SignedBeaconBlockSchema phase0BlockSchema =
-        spec.forMilestone(SpecMilestone.PHASE0).getSchemaDefinitions().getSignedBeaconBlockSchema();
-    final RpcContextCodec<Bytes, SignedBeaconBlock> noContextCodec =
-        RpcContextCodec.noop(phase0BlockSchema);
+    final RpcContextCodec<Bytes4, SignedBeaconBlock> forkDigestContextCodec =
+        RpcContextCodec.forkDigest(
+            spec, recentChainData, ForkDigestPayloadContext.SIGNED_BEACON_BLOCK);
 
     final SingleProtocolEth2RpcMethod<BeaconBlocksByRootRequestMessage, SignedBeaconBlock>
-        v1Method =
+        v2Method =
             new SingleProtocolEth2RpcMethod<>(
                 asyncRunner,
                 BeaconChainMethodIds.BEACON_BLOCKS_BY_ROOT,
-                1,
+                2,
                 rpcEncoding,
                 requestType,
                 expectResponseToRequest,
-                noContextCodec,
+                forkDigestContextCodec,
                 beaconBlocksByRootHandler,
                 peerLookup);
 
-    if (spec.isMilestoneSupported(SpecMilestone.ALTAIR)) {
-      final RpcContextCodec<Bytes4, SignedBeaconBlock> forkDigestContextCodec =
-          RpcContextCodec.forkDigest(
-              spec, recentChainData, ForkDigestPayloadContext.SIGNED_BEACON_BLOCK);
-
-      final SingleProtocolEth2RpcMethod<BeaconBlocksByRootRequestMessage, SignedBeaconBlock>
-          v2Method =
-              new SingleProtocolEth2RpcMethod<>(
-                  asyncRunner,
-                  BeaconChainMethodIds.BEACON_BLOCKS_BY_ROOT,
-                  2,
-                  rpcEncoding,
-                  requestType,
-                  expectResponseToRequest,
-                  forkDigestContextCodec,
-                  beaconBlocksByRootHandler,
-                  peerLookup);
-
-      return VersionedEth2RpcMethod.create(
-          rpcEncoding, requestType, expectResponseToRequest, List.of(v2Method, v1Method));
-    } else {
-      return v1Method;
-    }
+    return VersionedEth2RpcMethod.create(
+        rpcEncoding, requestType, expectResponseToRequest, List.of(v2Method));
   }
 
   private static Eth2RpcMethod<BeaconBlocksByRangeRequestMessage, SignedBeaconBlock>
@@ -260,56 +235,33 @@ public class BeaconChainMethods {
           final CombinedChainDataClient combinedChainDataClient,
           final PeerLookup peerLookup,
           final RpcEncoding rpcEncoding) {
-
     final BeaconBlocksByRangeMessageHandler beaconBlocksByRangeHandler =
         new BeaconBlocksByRangeMessageHandler(
             spec, metricsSystem, combinedChainDataClient, MAX_BLOCK_BY_RANGE_REQUEST_SIZE);
-    // V1 request only deal with Phase0 blocks
-    final SignedBeaconBlockSchema phase0BlockSchema =
-        spec.forMilestone(SpecMilestone.PHASE0).getSchemaDefinitions().getSignedBeaconBlockSchema();
-    final RpcContextCodec<?, SignedBeaconBlock> noContextCodec =
-        RpcContextCodec.noop(phase0BlockSchema);
 
     final BeaconBlocksByRangeRequestMessageSchema requestType =
         BeaconBlocksByRangeRequestMessage.SSZ_SCHEMA;
     final boolean expectResponseToRequest = true;
 
+    final RpcContextCodec<Bytes4, SignedBeaconBlock> forkDigestContextCodec =
+        RpcContextCodec.forkDigest(
+            spec, recentChainData, ForkDigestPayloadContext.SIGNED_BEACON_BLOCK);
+
     final SingleProtocolEth2RpcMethod<BeaconBlocksByRangeRequestMessage, SignedBeaconBlock>
-        v1Method =
+        v2Method =
             new SingleProtocolEth2RpcMethod<>(
                 asyncRunner,
                 BeaconChainMethodIds.BEACON_BLOCKS_BY_RANGE,
-                1,
+                2,
                 rpcEncoding,
                 requestType,
                 expectResponseToRequest,
-                noContextCodec,
+                forkDigestContextCodec,
                 beaconBlocksByRangeHandler,
                 peerLookup);
 
-    if (spec.isMilestoneSupported(SpecMilestone.ALTAIR)) {
-      final RpcContextCodec<Bytes4, SignedBeaconBlock> forkDigestContextCodec =
-          RpcContextCodec.forkDigest(
-              spec, recentChainData, ForkDigestPayloadContext.SIGNED_BEACON_BLOCK);
-
-      final SingleProtocolEth2RpcMethod<BeaconBlocksByRangeRequestMessage, SignedBeaconBlock>
-          v2Method =
-              new SingleProtocolEth2RpcMethod<>(
-                  asyncRunner,
-                  BeaconChainMethodIds.BEACON_BLOCKS_BY_RANGE,
-                  2,
-                  rpcEncoding,
-                  requestType,
-                  expectResponseToRequest,
-                  forkDigestContextCodec,
-                  beaconBlocksByRangeHandler,
-                  peerLookup);
-
-      return VersionedEth2RpcMethod.create(
-          rpcEncoding, requestType, expectResponseToRequest, List.of(v2Method, v1Method));
-    } else {
-      return v1Method;
-    }
+    return VersionedEth2RpcMethod.create(
+        rpcEncoding, requestType, expectResponseToRequest, List.of(v2Method));
   }
 
   private static Optional<