+++ services/chainstorage/src/main/java/tech/pegasys/teku/services/chainstorage/StorageService.java
@@ -49,14 +49,17 @@ public class StorageService extends Service implements StorageServiceFacade {
   private volatile Optional<BlockPruner> blockPruner = Optional.empty();
   private volatile Optional<BlobsSidecarPruner> blobsPruner = Optional.empty();
   private final boolean depositSnapshotStorageEnabled;
+  private final boolean blobsSidecarStorageCountersEnabled;
 
   public StorageService(
       final ServiceConfig serviceConfig,
       final StorageConfiguration storageConfiguration,
-      final boolean depositSnapshotStorageEnabled) {
+      final boolean depositSnapshotStorageEnabled,
+      final boolean blobsSidecarStorageCountersEnabled) {
     this.serviceConfig = serviceConfig;
     this.config = storageConfiguration;
     this.depositSnapshotStorageEnabled = depositSnapshotStorageEnabled;
+    this.blobsSidecarStorageCountersEnabled = blobsSidecarStorageCountersEnabled;
   }
 
   @Override
@@ -93,10 +96,12 @@ public class StorageService extends Service implements StorageServiceFacade {
                         new BlobsSidecarPruner(
                             config.getSpec(),
                             database,
+                            serviceConfig.getMetricsSystem(),
                             storagePrunerAsyncRunner,
                             serviceConfig.getTimeProvider(),
                             config.getBlobsPruningInterval(),
-                            config.getBlobsPruningLimit()));
+                            config.getBlobsPruningLimit(),
+                            blobsSidecarStorageCountersEnabled));
               }
               final EventChannels eventChannels = serviceConfig.getEventChannels();
               chainStorage = ChainStorage.create(database, config.getSpec());